requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
PyMySQL
